requirements.txt
# MOBSCAN v1.1.0 - Requirements
# Mobile Application Security Testing Framework

# Core Dependencies
python>=3.8

# Async Support
asyncio>=3.4.3
aiohttp>=3.8.0
aiofiles>=23.0.0

# Configuration & Data
PyYAML>=6.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# CLI & Output
click>=8.1.0
colorama>=0.4.6
rich>=13.0.0
tabulate>=0.9.0

# Reports
Jinja2>=3.1.0
markdown>=3.4.0
weasyprint>=59.0  # Para PDF
python-docx>=0.8.11  # Para DOCX

# Mobile Analysis
# APK Analysis
androguard>=3.4.0
# IPA Analysis (opcional)
# biplist>=1.0.3

# Network & Proxy
mitmproxy>=9.0.0
requests>=2.31.0
urllib3>=2.0.0

# Frida (Runtime Instrumentation)
frida>=16.0.0
frida-tools>=12.0.0

# Security & Crypto
cryptography>=41.0.0
pyOpenSSL>=23.0.0

# Vulnerability Database
# safety>=2.3.0
# requests para APIs de vulnerabilidades

# Code Analysis
regex>=2023.0.0
pyflakes>=3.0.0

# SBOM & SCA
cyclonedx-python-lib>=5.0.0
packageurl-python>=0.11.0

# Database (opcional)
# SQLAlchemy>=2.0.0

# Logging
loguru>=0.7.0

# Performance
multiprocessing-logging>=0.3.0

# Utils
tqdm>=4.66.0  # Progress bars
python-magic>=0.4.27  # File type detection
